issue 01, 02: renamed all additional letters and added them into the App
